Add Bulk API Job option to change concurrency mode
There are some options supported by the Bulk API that are missing from jsforce. This simple commit adds concurrencyMode, which allows users to toggle between Parallel (the default in Salesforce) and Serial.

From the Salesforce docs:

Parallel: Process batches in parallel mode. This is the default value.
Serial: Process batches in serial mode. Processing in parallel can cause database contention. When this is severe, the job may fail. If you're experiencing this issue, submit the job with serial concurrency mode. This guarantees that batches are processed one at a time. Note that using this option may significantly increase the processing time for a job.