Fix update-, create- and destoryParallel exception handler
The catch handlers for the update-, create- and destoryParallel methods in the connection base class make a call to _toRecordResult on this. In the context of the catch-handler this will not set to the instance of the connection. Instead of calling _toRecordResult on this it should be called on self which is already defined for all methods.