refactor: add type to type only imports and exports
This is to play nice with bundlers.
babel removes the type info from the types files since js does not care about them.
This has the effect of having an import with no export which bundlers do not like.
You can see this most clearly in the analysis.ts file which
is re exporting types from the types file.

https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-8.html#type-only-imports-and-export